requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pymongo>=4.0.0
motor>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
